Move formula calculators to their own file
