+++ server/src/controllers/attainable.ts
@@ -13,6 +13,7 @@ import UserAttainmentGrade from '../database/models/userAttainmentGrade';
 
 import { AttainableData, AttainableRequestData, Formula, FormulaParams } from '../types/attainable';
 import { ApiError } from '../types/error';
+import { CalculationResult, formulaChecker, FormulaNode, getFormula, Status } from '../types/formulas';
 import { idSchema } from '../types/general';
 import { HttpCode } from '../types/httpCode';
 import { findAttainableById, generateAttainableTag } from './utils/attainable';
@@ -235,92 +236,6 @@ export async function updateAttainable(req: Request, res: Response): Promise<voi
       }
     }
   });
-}
-
-enum Status {
-  Pass = 'pass',
-  Fail = 'fail',
-}
-
-interface CalculationResult {
-  status: Status;
-  points: number | undefined;
-}
-
-interface WeightedAssignmentParams {
-  min: number;
-  max: number;
-  weights: Array<number>;
-}
-
-type FormulaFunction = (subResults: Array<CalculationResult>) => Promise<CalculationResult>;
-type ParameterizedFormulaFunction = (parameters: any, subResults: Array<CalculationResult>) => Promise<CalculationResult>;
-interface FormulaNode {
-  validatedFormula: FormulaFunction;
-  subFormulaNodes: Array<FormulaNode>;
-};
-
-const formulasWithSchema: Map<Formula, [yup.AnyObjectSchema, ParameterizedFormulaFunction]> = new Map();
-formulasWithSchema.set(
-  Formula.Manual,
-  [
-    yup.object(),
-    async (_subGrades, _params) => { return { status: Status.Fail, points: undefined }; },
-  ]
-);
-
-async function calculatedWeightedAverage(
-  params: WeightedAssignmentParams,
-  subResults: Array<CalculationResult>
-): Promise<CalculationResult> {
-  const weighted: CalculationResult =
-    params.weights
-      .reduce(
-        (
-          acc: CalculationResult,
-          weight: number,
-          i: number,
-        ) => {
-          if (acc.status == Status.Fail || subResults[i].status == Status.Fail) {
-            return { points: undefined, status: Status.Fail };
-          }
-          return {
-            points: (acc.points ?? 0) + weight * (subResults[i].points ?? 0),
-            status: Status.Pass,
-          };
-        },
-        { status: Status.Pass, points: 0 }
-      );
-
-  return weighted;
-}
-
-formulasWithSchema.set(
-  Formula.WeightedAverage,
-  [
-    yup.object({
-      min: yup.number().required(),
-      max: yup.number().required(),
-      weights: yup.array(yup.number().required()).required(),
-    }),
-    calculatedWeightedAverage,
-  ]
-);
-
-const formulaChecker = yup.string().oneOf(Object.values(Formula)).required();
-
-async function validate<P extends FormulaParams>(
-  fn: (params: P, subPoints: Array<CalculationResult>) => Promise<CalculationResult>,
-  schema: yup.AnyObjectSchema,
-  params: unknown,
-): Promise<FormulaFunction> {
-  await schema.validate(params);
-  return (subGrades) => fn(params as P, subGrades);
-}
-
-function getFormula(name: Formula, params: FormulaParams) {
-  const formulaWithSchema = formulasWithSchema.get(name)!;
-  return validate(formulaWithSchema[1], formulaWithSchema[0], params);
 }
 
 async function calculate(tree: FormulaNode, presetPoints: Map<FormulaNode, number>): Promise<CalculationResult> {
